+++ android/src/main/kotlin/dev/msfjarvis/claw/android/ui/navigation/Destinations.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2021-2022 Harsh Shandilya.
+ * Copyright © 2021-2023 Harsh Shandilya.
  * Use of this source code is governed by an MIT-style
  * license that can be found in the LICENSE file or at
  * https://opensource.org/licenses/MIT.
@@ -29,6 +29,10 @@ sealed class Destinations {
   object User : Destinations() {
     const val placeholder = "{username}"
     override val route = "user/$placeholder"
+  }
+
+  object DataTransfer : Destinations() {
+    override val route: String = "datatransfer"
   }
 
   companion object {